feat: MainActivityの一部処理を担うDelegateを作成しました
